revert bridge loop extraction
