--- loraham_kiss_tnc/bridge_loop.h
@@ -1,8 +1,0 @@
-#ifndef BRIDGE_LOOP_H
-#define BRIDGE_LOOP_H
-
-#include "loraham_kiss_tnc.h"
-
-int bridge_loop_run(const lhkt_config_t *cfg, lhkt_stats_t *stats);
-
-#endif
