+++ loraham_kiss_tnc/build.sh
@@ -60,7 +60,6 @@ sources=(
   bridge_kiss.c
   bridge_rx.c
   bridge_loraham.c
-  bridge_loop.c
   bridge_runtime.c
   bridge_tx_queue.c
   cli.c
